refactor: Rename ui/presentation to ui/screens
